Be consistent with 'identifier' naming
Sometimes it's written as id or ID or IDS, but an id in JavaScript is `#id`, while an identifier for form validation can be either an `id`, `name` or `data-validate` value. To avoid confusing let's keep it all the same as how it's introduced in the beginning.

I'm I right that fieldName is actually limited to a `name` value? It's not easy to decipher that from the code...