feat(chore): added testing page for jsfiddle fullscreen
jsfiddle.net recently removed the feature to append a /show to any jsfiddle , which was then only showing the result of a jsfiddle. That was a very handy feature to review/test on mobile devices.

The only possibility now is to embed the jsfiddle into an iframe.

That said, this PR offers a little page which will display the result of any given jsfiddle in an iframe.
Maybe i was too lazy, but i could not find another public service which does just that.

I am aware, that codepen.io still offers that feature somehow by using /full instead of /pen in their URLs, but as we are using jsfiddle so much often i thought it's a good idea to still have this for jsfiddle as well.