fix(dropdown): incorrect format in example code of selector settings
The child combinator syntax '>' is encoded in the example of selector
settings and it's not readable for the user. The syntax is encoded into
HTML character as '&gt;' because the example code block is formatted as
HTML code block via 'data-type=html'.

The selector settings is for CSS and so it's formatted as CSS code block
by changing 'data-type=css'.