fix(emoji): fix anchor link generation
The section links on the right side are not working for emoji categories.

The emoji list is generated by javascript. However, this was done after the document was loaded and AFTER the central docs.js was auto generating the achor links. This PR changes the order of the execution, so by the time the anchor link routine seeks for proper segments, those are available at runtime now.