fix(build): clarify directory used for build
Recently added Fomantic UI to one of my projects. I believe this is the directory to be accessed for the build command, assuming the directory was not changed in semantic.json. I think the node_modules/fomantic-ui directory would work as well, but if you installed Fomantic using npx gulp install, you might want to start running the build command from semantic/ directory instead, especially if you want to make changes to it in the future.