fix(less): use cloudflare polyfill mirror
polyfill.io is out of service. We still need this as long as we support IE11 (until 2.10), as LESS uses JS functionality which does not work in IE11. ( See #308 )

This PR replaces the original URL by a cloudflare link as they mirrored the same functionality some time ago.
blog.cloudflare.com/polyfill-io-now-available-on-cdnjs-reduce-your-supply-chain-risk
blog.cloudflare.com/automatically-replacing-polyfill-io-links-with-cloudflares-mirror-for-a-safer-internet