feat(chore): iframe check, use let, remove IE console settings
The docs page is not intended to be run inside a foreign iframe.
Usually we would use the X-frame-options http header to prevent this, but as the docs run on github pages, it is not possible to set this header.
Therefore this PR shows a banner to tell the user about a possible clickjacking attempt

the PR also changes all var statements to let and removes old IE related code